Migrate ES RestClient to indy plugin
