+++ apm-agent-plugins/apm-es-restclient-plugin/apm-es-restclient-plugin-5_6/src/main/resources/META-INF/services/co.elastic.apm.agent.sdk.ElasticApmInstrumentation
@@ -1,2 +1,2 @@
-co.elastic.apm.agent.es.restclient.v5_6.ElasticsearchClientSyncInstrumentation
-co.elastic.apm.agent.es.restclient.v5_6.ElasticsearchClientAsyncInstrumentation
+co.elastic.apm.agent.esrestclient.v5_6.ElasticsearchClientSyncInstrumentation
+co.elastic.apm.agent.esrestclient.v5_6.ElasticsearchClientAsyncInstrumentation
